requirements.txt
fastapi
pydantic
uvicorn
jinja2
pymongo
python-dotenv
boto3
mysql.connector
